common: Remove layout slice and move selectors to grid component
