+++ packages/browser-sdk/src/lib/react/Grid/index.tsx
@@ -1,11 +1,10 @@
 import * as React from "react";
 
 import { VideoView, WherebyVideoElement } from "../VideoView";
-import { setAspectRatio, debounce } from "@whereby.com/core";
+import { debounce } from "@whereby.com/core";
 import { CellView, Bounds, Origin } from "./layout/types";
 import { VideoStageLayout } from "./VideoStageLayout";
 import { useGrid } from "./useGrid";
-import { useAppDispatch } from "../Provider/hooks";
 
 interface RenderCellViewProps {
     cellView: CellView;
@@ -79,11 +78,16 @@ interface GridProps {
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 function Grid({ renderParticipant }: GridProps) {
-    const dispatch = useAppDispatch();
     const gridRef = React.useRef<HTMLDivElement>(null);
 
-    const { cellViewsVideoGrid, cellViewsInPresentationGrid, cellViewsInSubgrid, videoStage, setContainerBounds } =
-        useGrid();
+    const {
+        cellViewsVideoGrid,
+        cellViewsInPresentationGrid,
+        cellViewsInSubgrid,
+        videoStage,
+        setContainerBounds,
+        setClientAspectRatios,
+    } = useGrid();
 
     const presentationGridContent = React.useMemo(
         () =>
@@ -91,7 +95,10 @@ function Grid({ renderParticipant }: GridProps) {
                 renderCellView({
                     cellView,
                     onSetClientAspectRatio: ({ aspectRatio, clientId }) =>
-                        dispatch(setAspectRatio({ clientId, aspectRatio })),
+                        setClientAspectRatios((prev) => ({
+                            ...prev,
+                            [clientId]: aspectRatio,
+                        })),
                 }),
             ),
         [cellViewsInPresentationGrid],
@@ -103,7 +110,10 @@ function Grid({ renderParticipant }: GridProps) {
                 renderCellView({
                     cellView,
                     onSetClientAspectRatio: ({ aspectRatio, clientId }) =>
-                        dispatch(setAspectRatio({ clientId, aspectRatio })),
+                        setClientAspectRatios((prev) => ({
+                            ...prev,
+                            [clientId]: aspectRatio,
+                        })),
                 }),
             ),
         [cellViewsVideoGrid],
@@ -115,7 +125,10 @@ function Grid({ renderParticipant }: GridProps) {
                 renderCellView({
                     cellView,
                     onSetClientAspectRatio: ({ aspectRatio, clientId }) =>
-                        dispatch(setAspectRatio({ clientId, aspectRatio })),
+                        setClientAspectRatios((prev) => ({
+                            ...prev,
+                            [clientId]: aspectRatio,
+                        })),
                 }),
             ),
         [cellViewsInSubgrid],
